requirements.txt
mysql-connector-python~=9.2.0
openpyxl
pandas~=2.2.3
Flask~=3.1.0
